+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/misc/QuadConsumer.java
@@ -1,0 +1,6 @@
+package net.mehvahdjukaar.moonlight.api.misc;
+
+@FunctionalInterface
+public interface QuadConsumer<K, V, S, T> {
+    void accept(K k, V v, S s, T t);
+}