+++ fabric/src/main/java/net/mehvahdjukaar/moonlight/fabric/MoonlightFabric.java
@@ -7,7 +7,6 @@ import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerWorldEvents;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
 import net.mehvahdjukaar.moonlight.api.fluids.SoftFluidRegistry;
-import net.mehvahdjukaar.moonlight.api.platform.RegHelper;
 import net.mehvahdjukaar.moonlight.api.platform.configs.fabric.FabricConfigSpec;
 import net.mehvahdjukaar.moonlight.api.platform.fabric.RegHelperImpl;
 import net.mehvahdjukaar.moonlight.api.platform.network.NetworkDir;
@@ -43,7 +42,7 @@ public class MoonlightFabric implements ModInitializer, DedicatedServerModInitia
     static void commonSetup() {
         RegHelperImpl.lateRegisterEntries();
         FabricConfigSpec.loadAllConfigs();
-        FabricSetupCallbacks.COMMON_SETUP.forEach(Runnable::run);
+        MLFabricSetupCallbacks.COMMON_SETUP.forEach(Runnable::run);
     }
 
     public static MinecraftServer currentServer;