+++ changelog.md
@@ -1,1 +1,1 @@
-* fixed fabric version
+* added creative tabs helpers